[#49] Specify which version of the tool was actually downloaded (#60)
This resolves #49

I took a stab at implementing this but Im not sure if my approach is the right
way to go. Because the list of tools is already known before actually calling
the github api the length of these names is also known. This is not that case
with the `tag_name`. The tags from the example file are all less than 7 long but
that does not guarantee other repos.

I could try and check what tags are available for all tools before actually
downloading them. Can I get your input on this?


```bash
$ cargo run -- --config tools.toml sync
    Finished dev [unoptimized + debuginfo] target(s) in 0.02s
     Running `target/debug/tool --config tools.toml sync`
✅  bat        v0.21.0  Completed!
✅  difftastic 0.34.0   Completed!
✅  exa        v0.10.1  Completed!
✅  fd         v8.4.0   Completed!
✅  ripgrep    13.0.0   Completed!
```